+++ main/pointctl.go
@@ -2,10 +2,11 @@ package main
 
 import (
 	"github.com/lightstar-dev/openlan-go/point"
+	"github.com/lightstar-dev/openlan-go/point/models"
 )
 
 func main() {
-	c := point.NewConfig()
+	c := models.NewConfig()
 	p := point.NewCommand(c)
 	p.Start()
 	p.Loop()
