+++ resource/.point.json
@@ -1,0 +1,5 @@
+{
+  "VsAddr": "ali.openlan.net",
+  "VsAuth": "yunex:yunex@123",
+  "IfAddr": "192.168.10.31/24"
+}
