--- resource/run.sh
@@ -1,2 +1,0 @@
-nohup ./resources/ope -br br-olan &
-nohup ./resources/cpe -br br-olan &
